Move scripts/helper.py to thai2transformers/helper
